Integrate remacs_sys as a module in the remacs crate. (#718)
Adjust README.md.  Merge build.rs.  Rename other files.
::remacs_sys::libm becomes ::libm.
