+++ Makefile.in
@@ -1173,7 +1173,6 @@ check-declare:
 
 rustfmt:
 	(cd rust_src && cargo fmt -- --write-mode overwrite)
-	(cd rust_src/remacs-sys && cargo fmt -- --write-mode overwrite)
 	(cd rust_src/remacs-lib && cargo fmt -- --write-mode overwrite)
 	(cd rust_src/remacs-util && cargo fmt -- --write-mode overwrite)
 	(cd rust_src/remacs-macros && cargo fmt -- --write-mode overwrite lib.rs)
