+++ rust_src/Cargo.toml
@@ -11,7 +11,6 @@ build = "build.rs"
 [dependencies]
 remacs-lib = { version = "0.1.0", path = "remacs-lib/" }
 remacs-macros = { version = "0.1.0", path = "remacs-macros" }
-remacs-sys = { version = "0.1.0", path = "remacs-sys/" }
 base64 = "0.9"
 clippy = { version = "*", optional = true }
 lazy_static = "0.2.2"
