+++ rust_src/src/floatfns.rs
@@ -5,11 +5,11 @@ use std::mem;
 
 use libc;
 
+use libm;
 use remacs_macros::lisp_fn;
 use remacs_sys::{EmacsDouble, EmacsInt, EmacsUint, Lisp_Object, MOST_NEGATIVE_FIXNUM,
                  MOST_POSITIVE_FIXNUM};
 use remacs_sys::{Qarith_error, Qinteger_or_marker_p, Qnumberp, Qrange_error};
-use remacs_sys::libm;
 
 use lisp::{LispNumber, LispObject};
 use lisp::defsubr;
