+++ rust_src/src/remacs_sys.rs
@@ -1,8 +1,4 @@
-#![cfg_attr(feature = "clippy", feature(plugin))]
-#![cfg_attr(feature = "clippy", plugin(clippy))]
-#![feature(const_size_of)]
-#![feature(repr_transparent)]
-#![allow(non_camel_case_types, non_snake_case, non_upper_case_globals)]
+#![allow(unused)]
 
 //! This module contains all FFI declarations.
 //!
@@ -20,8 +16,7 @@
 //! - `BoolBF`
 
 extern crate libc;
-
-pub mod libm;
+extern crate std;
 
 use libc::{c_char, c_double, c_float, c_int, c_short, c_uchar, c_void, intmax_t, off_t, ptrdiff_t,
            size_t, time_t, timespec};
